+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Fixed
 
-- None
+- Mark UIImage extension initializers as available for watchOS (By [Tomas Franzén](https://github.com/tomasf))
 
 ## [2.1.1] - 2020-12-13